Share drawer on mobile too
