+++ packages/mobile/src/components/usdc-manual-transfer-drawer/USDCManualTransferDrawer.tsx
@@ -1,53 +1,14 @@
-import { useCallback, useMemo } from 'react'
-
-import {
-  Name,
-  isContentPurchaseInProgress,
-  purchaseContentSelectors,
-  useCreateUserbankIfNeeded,
-  useUSDCBalance,
-  useUSDCManualTransferModal
-} from '@audius/common'
-import { USDC } from '@audius/fixed-decimal'
-import Clipboard from '@react-native-clipboard/clipboard'
-import BN from 'bn.js'
+import { useUSDCManualTransferModal } from '@audius/common'
 import { View } from 'react-native'
-import QRCode from 'react-qr-code'
-import { useDispatch, useSelector } from 'react-redux'
-import { useAsync } from 'react-use'
 
-import IconError from 'app/assets/images/iconError.svg'
-import LogoUSDC from 'app/assets/images/logoUSDC.svg'
-import { Button, Text, useLink } from 'app/components/core'
+import { Text } from 'app/components/core'
 import Drawer from 'app/components/drawer'
-import { useToast } from 'app/hooks/useToast'
-import { make, track, track as trackEvent } from 'app/services/analytics'
-import { audiusBackendInstance } from 'app/services/audius-backend-instance'
-import { getUSDCUserBank } from 'app/services/buyCrypto'
 import { flexRowCentered, makeStyles } from 'app/styles'
-import { spacing } from 'app/styles/spacing'
-import type { AllEvents } from 'app/types/analytics'
-import { useThemeColors } from 'app/utils/theme'
 
-import { AddressTile } from '../core/AddressTile'
-
-const { getPurchaseContentFlowStage, getPurchaseContentError } =
-  purchaseContentSelectors
-
-const USDCLearnMore =
-  'https://support.audius.co/help/Understanding-USDC-on-Audius'
+import { USDCManualTransfer } from '../usdc-manual-transfer/USDCManualTransfer'
 
 const messages = {
-  title: 'Manual Crypto Transfer',
-  explainer:
-    'Add funds by sending Solana based (SPL) USDC to your Audius account.',
-  hint: 'Use caution to avoid errors and lost funds.',
-  copy: 'Copy Wallet Address',
-  goBack: 'Go Back',
-  learnMore: 'Learn More',
-  copied: 'Copied to Clipboard!',
-  usdcBalance: 'USDC Balance',
-  buy: (amount: string) => `Buy ${amount}`
+  title: 'Manual Crypto Transfer'
 }
 
 const useStyles = makeStyles(({ spacing, palette, typography }) => ({
@@ -63,102 +24,12 @@ const useStyles = makeStyles(({ spacing, palette, typography }) => ({
     paddingBottom: spacing(4),
     borderBottomWidth: 1,
     borderBottomColor: palette.neutralLight8
-  },
-  disclaimerContainer: {
-    display: 'flex',
-    alignItems: 'flex-start',
-    flexDirection: 'row',
-    paddingHorizontal: spacing(4),
-    paddingVertical: spacing(3),
-    backgroundColor: palette.backgroundSurface2,
-    borderColor: palette.borderStrong,
-    borderWidth: 1,
-    borderRadius: spacing(2),
-    gap: spacing(4)
-  },
-  disclaimer: {
-    lineHeight: typography.fontSize.medium * 1.25
-  },
-  icon: {
-    marginTop: spacing(2)
-  },
-  buttonContainer: {
-    gap: spacing(2)
-  },
-  learnMore: {
-    textDecorationLine: 'underline'
-  },
-  explainer: {
-    textAlign: 'left',
-    lineHeight: typography.fontSize.medium * 1.25
-  },
-  hintContainer: {
-    gap: spacing(3),
-    flexShrink: 1
-  },
-  shrink: {
-    flexShrink: 1
-  },
-  qr: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 10
   }
 }))
 
 export const USDCManualTransferDrawer = () => {
   const styles = useStyles()
-  const dispatch = useDispatch()
-  const { toast } = useToast()
-  const { isOpen, onClose, onClosed, data } = useUSDCManualTransferModal()
-  const { amount: amountInCents, onSuccessAction } = data ?? {}
-  const { onPress: onPressLearnMore } = useLink(USDCLearnMore)
-  const { neutral, specialLightGreen } = useThemeColors()
-
-  const stage = useSelector(getPurchaseContentFlowStage)
-  const error = useSelector(getPurchaseContentError)
-  const isUnlocking = !error && isContentPurchaseInProgress(stage)
-  const { data: balanceBN } = useUSDCBalance({
-    isPolling: true,
-    pollingInterval: 1000
-  })
-  const balance = USDC(balanceBN ?? new BN(0)).value
-  const amount = USDC((amountInCents ?? 0) / 100).value
-  const isBuyButtonDisabled = isUnlocking || balance < amount
-
-  useCreateUserbankIfNeeded({
-    recordAnalytics: track,
-    audiusBackendInstance,
-    mint: 'usdc'
-  })
-
-  const { value: USDCUserBank } = useAsync(async () => {
-    const USDCUserBankPubKey = await getUSDCUserBank()
-    return USDCUserBankPubKey?.toString() ?? ''
-  })
-
-  const analytics: AllEvents = useMemo(
-    () => ({
-      eventName: Name.PURCHASE_CONTENT_USDC_USER_BANK_COPIED,
-      address: USDCUserBank ?? ''
-    }),
-    [USDCUserBank]
-  )
-
-  const handleConfirmPress = useCallback(() => {
-    Clipboard.setString(USDCUserBank ?? '')
-    toast({ content: messages.copied, type: 'info' })
-    trackEvent(make(analytics))
-  }, [USDCUserBank, analytics, toast])
-
-  const handleLearnMorePress = useCallback(() => {
-    onPressLearnMore()
-  }, [onPressLearnMore])
-
-  const handleBuyPress = useCallback(() => {
-    if (onSuccessAction) dispatch(onSuccessAction)
-    onClose()
-  }, [dispatch, onClose, onSuccessAction])
+  const { isOpen, onClose, onClosed } = useUSDCManualTransferModal()
 
   return (
     <Drawer isOpen={isOpen} onClose={onClose} onClosed={onClosed}>
@@ -174,82 +45,7 @@ export const USDCManualTransferDrawer = () => {
             {messages.title}
           </Text>
         </View>
-        <Text style={styles.explainer}>{messages.explainer}</Text>
-        <View style={styles.qr}>
-          {USDCUserBank ? (
-            <QRCode size={160} style={styles.qr} value={USDCUserBank} />
-          ) : null}
-        </View>
-        <AddressTile
-          title={messages.usdcBalance}
-          address={USDCUserBank ?? ''}
-          left={<LogoUSDC height={spacing(6)} />}
-          analytics={analytics}
-          balance={USDC(balanceBN ?? new BN(0)).toLocaleString('en-US', {
-            roundingMode: 'floor',
-            minimumFractionDigits: 2,
-            maximumFractionDigits: 2
-          })}
-        />
-        <View style={styles.disclaimerContainer}>
-          <IconError
-            width={spacing(6)}
-            height={spacing(6)}
-            fill={neutral}
-            style={styles.icon}
-          />
-          <View style={styles.hintContainer}>
-            <View style={styles.shrink}>
-              <Text style={styles.disclaimer}>{messages.hint}</Text>
-            </View>
-            <Text
-              style={styles.learnMore}
-              color='primary'
-              onPress={handleLearnMorePress}
-            >
-              {messages.learnMore}
-            </Text>
-          </View>
-        </View>
-        <View style={styles.buttonContainer}>
-          {amountInCents === undefined ? (
-            <>
-              <Button
-                title={messages.copy}
-                onPress={handleConfirmPress}
-                variant='primary'
-                size='large'
-                fullWidth
-              />
-              <Button
-                title={messages.goBack}
-                onPress={onClose}
-                variant='common'
-                size='large'
-                fullWidth
-              />
-            </>
-          ) : (
-            <>
-              <Button
-                title={messages.goBack}
-                onPress={onClose}
-                variant='common'
-                size='large'
-                fullWidth
-              />
-              <Button
-                title={messages.buy(USDC(amount).ceil(2).toFixed(2))}
-                onPress={handleBuyPress}
-                variant='primary'
-                color={specialLightGreen}
-                size='large'
-                disabled={isBuyButtonDisabled}
-                fullWidth
-              />
-            </>
-          )}
-        </View>
+        <USDCManualTransfer onClose={onClose} />
       </View>
     </Drawer>
   )
