+++ packages/mobile/src/components/usdc-manual-transfer/index.ts
@@ -1,0 +1,1 @@
+export { USDCManualTransfer } from './USDCManualTransfer'
