+++ packages/web/src/components/premium-content-purchase-modal/components/PurchaseContentFormFooter.tsx
@@ -63,7 +63,6 @@ type PurchaseContentFormFooterProps = Pick<
 > & {
   track: PurchasableTrackMetadata
   onViewTrackClicked: () => void
-  onSubmit: () => void
 }
 
 export const PurchaseContentFormFooter = ({
